examples/authorization-for-rag/langchain-examples/pyproject.toml
[tool.poetry]
name = "langchain-examples"
version = "0.1.0"
description = ""
authors = ["Jose F. Romaniello <jose@okta.com>", "Javier Centurion <javier.centurion@okta.com>"]
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.scripts]
langchain-examples-scripts = "langchain_rag.main:main"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.11"
langchain-openai = "^0.2.12"
faiss-cpu = "^1.9.0.post1"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.11"
langchain-auth0-ai = {path = "../../../packages/langchain-auth0-ai", develop = true}
openfga-sdk = "^0.8.1"
termcolor = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
